requirements.txt
urllib3==1.26.18
requests==2.32.3
pandas==1.3.2
python-dotenv==0.19.0
backports.zoneinfo==0.2.1
